+++ CHANGELOG.md
@@ -3,6 +3,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
  * Bumped runtimeVersion to `2020.07.03`.
+ * `search` supports order by `points`.
 
 ## `20200702t124059-all`
  * Bumped runtimeVersion to `2020.07.02`.